requirements.txt
langchain
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
langchain_community
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette